// src/thrift/api_types.d.ts
declare enum TransactionType {
  TT_Transfer = 0,
  TT_ContractDeploy = 1,
  TT_ContractCall = 2,
  TT_ContractState = 3,
  TT_ContractReplenish = 4,
  TT_TokenDeploy = 5,
  TT_TokenTransfer = 6,
  TT_Delegation = 7,
  TT_RevokeDelegation = 8,
  TT_Hold = 9,
  TT_Release = 10,
  TT_CancelHold = 11,
  TT_DelayedTransfer = 12,
  TT_UpdateBootstrapList = 13,
  TT_UpdateSettings = 14,
  TT_Malformed = 15,
  TT_ContractEmitted = 16,
  TT_Utility = 17,
  TT_Other = 18,
}
// ...
declare class AmountCommission {
    public commission?: number;

      constructor(args?: { commission?: number; });
  }
// ...
declare class SmartContractInvocation {
    public method: string;
    public params: general_ttypes.Variant[];
    public usedContracts: string[];
    public forgetNewState: boolean;
    public smartContractDeploy?: SmartContractDeploy;
    public version?: number;

      constructor(args?: { method: string; params: general_ttypes.Variant[]; usedContracts: string[]; forgetNewState: boolean; smartContractDeploy?: SmartContractDeploy; version?: number; });
  }

declare class TransactionId {
    public poolSeq: Int64;
    public index: number;

      constructor(args?: { poolSeq: Int64; index: number; });
  }
// ...
declare class ExtraFee {
    public sum: general_ttypes.Amount;
    public comment: string;
    public transactionId: TransactionId;

      constructor(args?: { sum: general_ttypes.Amount; comment: string; transactionId: TransactionId; });
  }

declare class SmartStateTransInfo {
    public success: boolean;
    public executionFee: general_ttypes.Amount;
    public returnValue?: general_ttypes.Variant;
    public startTransaction: TransactionId;

      constructor(args?: { success: boolean; executionFee: general_ttypes.Amount; returnValue?: general_ttypes.Variant; startTransaction: TransactionId; });
  }

declare class SmartTransInfo {
    public v_tokenDeploy?: TokenDeployTransInfo;
    public v_tokenTransfer?: TokenTransferTransInfo;
    public v_smartDeploy?: SmartDeployTransInfo;
    public v_smartExecution?: SmartExecutionTransInfo;
    public v_smartState?: SmartStateTransInfo;

      constructor(args?: { v_tokenDeploy?: TokenDeployTransInfo; v_tokenTransfer?: TokenTransferTransInfo; v_smartDeploy?: SmartDeployTransInfo; v_smartExecution?: SmartExecutionTransInfo; v_smartState?: SmartStateTransInfo; });
  }

declare class Transaction {
    public id: Int64;
    public source: string;
    public target: string;
    public amount: general_ttypes.Amount;
    public balance: general_ttypes.Amount;
    public currency: any;
    public signature: string;
    public smartContract?: SmartContractInvocation;
    public fee: AmountCommission;
    public timeCreation: Int64;
    public userFields?: string;
    public type: TransactionType;
    public smartInfo?: SmartTransInfo;
    public extraFee?: ExtraFee[];
    public poolNumber: Int64;
    public usedContracts?: string[];

      constructor(args?: { id: Int64; source: string; target: string; amount: general_ttypes.Amount; balance: general_ttypes.Amount; currency: any; signature: string; smartContract?: SmartContractInvocation; fee: AmountCommission; timeCreation: Int64; userFields?: string; type: TransactionType; smartInfo?: SmartTransInfo; extraFee?: ExtraFee[]; poolNumber: Int64; usedContracts?: string[]; });
  }

declare class SealedTransaction {
    public id: TransactionId;
    public trxn: Transaction;

      constructor(args?: { id: TransactionId; trxn: Transaction; });
  }
// ...
declare class DelegatedItem {
    public wallet: string;
    public sum: general_ttypes.Amount;
    public validUntil?: Int64;

      constructor(args?: { wallet: string; sum: general_ttypes.Amount; validUntil?: Int64; });
  }

declare class Delegated {
    public incoming: general_ttypes.Amount;
    public outgoing: general_ttypes.Amount;
    public donors?: DelegatedItem[];
    public recipients?: DelegatedItem[];

      constructor(args?: { incoming: general_ttypes.Amount; outgoing: general_ttypes.Amount; donors?: DelegatedItem[]; recipients?: DelegatedItem[]; });
  }

declare class WalletData {
    public walletId: number;
    public balance: general_ttypes.Amount;
    public lastTransactionId: Int64;
    public delegated?: Delegated;

      constructor(args?: { walletId: number; balance: general_ttypes.Amount; lastTransactionId: Int64; delegated?: Delegated; });
  }